# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContextOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
  /agents/{id}/model-policy:
    put:
      operationId: setAgentModelPolicy
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentModelPolicyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentModelPolicy"
  /model-endpoints:
    get:
      operationId: listModelEndpoints
      tags: [model-endpoints]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelEndpoint"
    post:
      operationId: createModelEndpoint
      tags: [model-endpoints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelEndpointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelEndpoint"
  /model-endpoints/{id}:
    get:
      operationId: getModelEndpoint
      tags: [model-endpoints]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelEndpoint"
    patch:
      operationId: updateModelEndpoint
      tags: [model-endpoints]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelEndpointUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelEndpoint"
    delete:
      operationId: deleteModelEndpoint
      tags: [model-endpoints]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
  /model-endpoints/{id}/test:
    post:
      operationId: testModelEndpoint
      tags: [model-endpoints]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelTestResult"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    Agent:
      type: object
      required: [id, name, role, contextPolicy, isActive, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        description: { type: ["string", "null"] }
        systemPrompt: { type: ["string", "null"] }
        capabilityScope: { type: ["array", "null"], items: { type: string } }
        contextPolicy: { type: string }
        exposeAsCapabilityProvider: { type: boolean }
        isActive: { type: boolean }
        modelPolicy: { $ref: "#/components/schemas/AgentModelPolicy" }
        createdAt: { type: string, format: date-time }
    AgentInput:
      type: object
      required: [name, role]
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string }
        description: { type: string }
        systemPrompt: { type: string }
        capabilityScope: { type: array, items: { type: string } }
        contextPolicy: { type: string }
        exposeAsCapabilityProvider: { type: boolean }
    AgentUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        systemPrompt: { type: string }
        capabilityScope: { type: array, items: { type: string } }
        contextPolicy: { type: string }
        isActive: { type: boolean }
        exposeAsCapabilityProvider: { type: boolean }
    AgentModelPolicy:
      type: object
      required: [id, agentId]
      properties:
        id: { type: string }
        agentId: { type: string }
        primaryEndpointId: { type: ["string", "null"] }
        fallbackEndpointId: { type: ["string", "null"] }
        temperature: { type: integer }
        maxTokens: { type: integer }
    AgentModelPolicyInput:
      type: object
      properties:
        primaryEndpointId: { type: string }
        fallbackEndpointId: { type: string }
        temperature: { type: integer }
        maxTokens: { type: integer }
...
    ModelEndpoint:
      type: object
      required: [id, name, providerType, modelName, status, isDefault, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        providerType: { type: string }
        baseUrl: { type: ["string", "null"] }
        host: { type: ["string", "null"] }
        port: { type: ["integer", "null"] }
        modelName: { type: string }
        apiKeyRef: { type: ["string", "null"] }
        apiKeyMasked: { type: ["string", "null"] }
        organization: { type: ["string", "null"] }
        deployment: { type: ["string", "null"] }
        requestTimeoutMs: { type: integer }
        maxRetries: { type: integer }
        isDefault: { type: boolean }
        status: { type: string }
        lastTestedAt: { type: ["string", "null"], format: date-time }
        lastTestResult: { $ref: "#/components/schemas/JsonObject" }
        createdAt: { type: string, format: date-time }
    ModelEndpointInput:
      type: object
      required: [name, providerType, modelName]
      properties:
        name: { type: string, minLength: 1 }
        providerType: { type: string }
        baseUrl: { type: string }
        host: { type: string }
        port: { type: integer }
        modelName: { type: string, minLength: 1 }
        apiKey: { type: string }
        organization: { type: string }
        deployment: { type: string }
        requestTimeoutMs: { type: integer }
        maxRetries: { type: integer }
        isDefault: { type: boolean }
    ModelEndpointUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        baseUrl: { type: string }
        host: { type: string }
        port: { type: integer }
        modelName: { type: string, minLength: 1 }
        apiKey: { type: string }
        organization: { type: string }
        deployment: { type: string }
        requestTimeoutMs: { type: integer }
        maxRetries: { type: integer }
        isDefault: { type: boolean }
        status: { type: string }
    ModelTestResult:
      type: object
      required: [ok, latencyMs, message]
      properties:
        ok: { type: boolean }
        latencyMs: { type: integer }
        message: { type: string }
        model: { type: ["string", "null"] }
        promptTokens: { type: ["integer", "null"] }
        completionTokens: { type: ["integer", "null"] }
        totalTokens: { type: ["integer", "null"] }
        usedStub: { type: boolean }
...