# lib/api-zod/src/generated/types/modelEndpoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContextOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JsonObject } from './jsonObject';

export interface ModelEndpoint {
  id: string;
  name: string;
  providerType: string;
  /** @nullable */
  baseUrl?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  port?: number | null;
  modelName: string;
  /** @nullable */
  apiKeyRef?: string | null;
  /** @nullable */
  apiKeyMasked?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  deployment?: string | null;
  requestTimeoutMs?: number;
  maxRetries?: number;
  isDefault: boolean;
  status: string;
  /** @nullable */
  lastTestedAt?: Date | null;
  lastTestResult?: JsonObject;
  createdAt: Date;
}