# lib/api-zod/src/generated/types/modelEndpointUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContextOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModelEndpointUpdate {
  /** @minLength 1 */
  name?: string;
  baseUrl?: string;
  host?: string;
  port?: number;
  /** @minLength 1 */
  modelName?: string;
  apiKey?: string;
  organization?: string;
  deployment?: string;
  requestTimeoutMs?: number;
  maxRetries?: number;
  isDefault?: boolean;
  status?: string;
}